v0.30 修复工具链恶性 BUG + 代码审查修复 + 版本单一来源
- BUG-031/032：cc500 文件槽泄漏（good2.c 二次编译失败）+ 自编译产物静默丢 argv；
  修复 + tests/repro_bugs.sh 双断言（good 源编译 OK -> 坏源 FAIL -> 同源再编译必须成功；
  ccboot 产 P1 -> exec /out.elf -> /out2.elf 必须被创建）
- BUG-033（代码审查 P0）：map_page_in 页表帧 OOM 未检查 -> 写物理 0 破坏低 4KB 且
  pf_handler 假增长死循环；改返回 int，pf_handler 失败转 STACK_BOOM 并释放已分配帧
- BUG-034（代码审查 P2）：kb 行缓冲 line_ready 期间仍追加输入致两行合并；仅
  !line_ready 才入缓冲，test_kb 补用例 13
- BUG-035（代码审查 P1，=OBS-002）：fork_frames[24] 硬编码改 kmalloc 动态数组，
  sched_fork 先数页数再按需分配、退出 kfree
- BUG-036（v0.30 内部回归）：L-5 误删 -Wno-int-conversion 只留 -w，GCC 14 下
  -Wint-conversion 是 permerror 硬错误、-w 压不住 -> make EXIT=2；恢复显式
  -Wno-int-conversion 与 -w 并存（GCC 8+ 均有效）
- 代码审查：sys_wait 缩进错位统一；fs_write 短写语义注释；kfree 合并 O(n²) 注释；
  e1000_tx 忙等 ISR 上下文注释
- 版本单一来源 src/version.h（评估 L-4）：启动横幅/shell banner/motd/两处回归断言统一；
  Makefile cc500 豁免精简为 -Wno-int-conversion -w（评估 L-5，含 BUG-036 修复）
- 文档回补：bugs.md（BUG-029~036、OBS-002 标记修复）、changelog v0.27c~v0.30、
  roadmap（fork_frames 动态化落地）、design（测试层数/内存布局核对）、
  README（根版本矩阵补 v0.29/v0.30 + 依赖 python3/socat；v2-c-kernel/README 五层/6 checks）

+++ v2-c-kernel/src/app/abuse.c
@@ -40,7 +40,7 @@ void app_main(int argc, char **argv) {
     report("print@0xFFFFFFF0",     syscall3(SYS_PRINT, 0xFFFFFFF0u, 0, 0));
 
     /* 文件读写缓冲指针：先打开一个合法文件，再用内核地址当缓冲 -> 应被拒 */
-    if (syscall3(SYS_FS_CREATE, (uint32_t)"/abuse.tmp", 0, 0) < 0) {
+    if ((int)syscall3(SYS_FS_CREATE, (uint32_t)"/abuse.tmp", 0, 0) < 0) {
         sys_print("[abuse] create /abuse.tmp failed\n");
     } else if (syscall3(SYS_FS_OPEN, 1, (uint32_t)"/abuse.tmp", 1) != 0) {
         sys_print("[abuse] open /abuse.tmp failed\n");
@@ -51,7 +51,7 @@ void app_main(int argc, char **argv) {
     }
 
     /* 合法路径应正常（不误伤）：建 -> 写 -> 删一个文件 */
-    if (syscall3(SYS_FS_CREATE, (uint32_t)"/ok.tmp", 0, 0) < 0 ||
+    if ((int)syscall3(SYS_FS_CREATE, (uint32_t)"/ok.tmp", 0, 0) < 0 ||
         syscall3(SYS_FS_OPEN, 1, (uint32_t)"/ok.tmp", 1) != 0) {
         sys_print("[abuse] valid-path FAIL\n");
         fail++;
